+++ dev/main-vite/src/main.ts
@@ -2,8 +2,8 @@ import {createApp} from 'vue'
 import App from './App.vue'
 // @ts-ignore
 import microApp from '@micro-zoe/micro-app'
-import ElementUI from 'element-plus'
-import 'element-plus/lib/theme-chalk/index.css';
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
 import router from './router';
 
 microApp.start({
@@ -22,47 +22,6 @@ microApp.start({
     },
     error() {
       console.log('error 全局监听')
-    }
-  },
-  plugins: {
-    modules: {
-      react16: [{
-        loader(code: string, url: string) {
-          if (code.indexOf('sockjs-node') > -1) {
-            console.log('react16插件', url)
-            code = code.replace('window.location.port', '3001')
-          }
-          return code
-        }
-      }],
-      react162: [{
-        loader(code: string, url: string) {
-          if (process.env.NODE_ENV === 'development' && code.indexOf('sockjs-node') > -1) {
-            console.log('react16插件', url)
-            code = code.replace('window.location.port', '3001')
-          }
-          return code
-        }
-      }],
-      react17: [{
-        loader(code: string, url: string) {
-          if (process.env.NODE_ENV === 'development' && code.indexOf('sockjs-node') > -1) {
-            console.log('react17插件', url)
-            code = code.replace('window.location.port', '3002')
-          }
-          return code
-        }
-      }],
-      vite: [{
-        loader(code: string) {
-          if (process.env.NODE_ENV === 'development') {
-            code = code.replace(/(from|import)(\s*['"])(\/micro-app\/vite\/)/g, (all) => {
-              return all.replace('/micro-app/vite/', 'http://localhost:7001/micro-app/vite/')
-            })
-          }
-          return code
-        }
-      }]
     }
   },
   /**
@@ -91,4 +50,4 @@ microApp.start({
   }
 })
 
-createApp(App).use(router).use(ElementUI).mount('#app')
+createApp(App).use(router).use(ElementPlus).mount('#app')