+++ src/constants.ts
@@ -58,18 +58,36 @@ export enum MicroAppConfig {
   FIBER = 'fiber',
 }
 
+/**
+ * global key must be static key, they can not rewrite
+ * e.g.
+ * window.Promise = newValue
+ * new Promise ==> still get old value, not newValue, because they are cached by top function
+ * NOTE:
+ * 1. Do not add fetch, XMLHttpRequest, EventSource
+ */
+export const GLOBAL_CACHED_KEY = 'window,self,globalThis,document,Document,Array,Object,String,Boolean,Math,Number,Symbol,Date,Function,Proxy,WeakMap,WeakSet,Set,Map,Reflect,Element,Node,RegExp,Error,TypeError,JSON,isNaN,parseFloat,parseInt,performance,console,decodeURI,encodeURI,decodeURIComponent,encodeURIComponent,navigator,undefined,location,history'
+
 // prefetch level
 export const PREFETCH_LEVEL: number[] = [1, 2, 3]
 
-// memory router constants
-// default mode, child router info will sync to browser url
-export const DEFAULT_ROUTER_MODE = 'search'
 /**
- * render base on browser url, and location.origin location.href point to base app
- * equal to disable-memory-router
+ * memory router modes
  * NOTE:
  *  1. The only difference between native and native-scope is location.origin, in native-scope mode location.origin point to child app
+ *  2. native mode equal to disable-memory-router
 */
+// 临时注释，1.0版本放开，默认模式切换为state
+// // default mode, sync child app router info to history.state
+// export const DEFAULT_ROUTER_MODE = 'state'
+// // sync child app router info to browser url as search
+// export const ROUTER_MODE_SEARCH = 'search'
+
+// 临时放开，1.0版本去除
+export const ROUTER_MODE_STATE = 'state'
+export const DEFAULT_ROUTER_MODE = 'search'
+
+// render base on browser url, and location.origin location.href point to base app
 export const ROUTER_MODE_NATIVE = 'native'
 // render base on browser url, but location.origin location.href point to child app
 export const ROUTER_MODE_NATIVE_SCOPE = 'native-scope'
@@ -77,6 +95,7 @@ export const ROUTER_MODE_NATIVE_SCOPE = 'native-scope'
 export const ROUTER_MODE_PURE = 'pure'
 export const ROUTER_MODE_LIST: string[] = [
   DEFAULT_ROUTER_MODE,
+  ROUTER_MODE_STATE,
   ROUTER_MODE_NATIVE,
   ROUTER_MODE_NATIVE_SCOPE,
   ROUTER_MODE_PURE,
@@ -142,12 +161,10 @@ export const GLOBAL_KEY_TO_WINDOW: Array<PropertyKey> = [
 
 export const RAW_GLOBAL_TARGET: Array<PropertyKey> = ['rawWindow', 'rawDocument']
 
-/**
- * global key must be static key, they can not rewrite
- * e.g.
- * window.Promise = newValue
- * new Promise ==> still get old value, not newValue, because they are cached by top function
- * NOTE:
- * 1. Do not add fetch, XMLHttpRequest, EventSource
- */
-export const GLOBAL_CACHED_KEY = 'window,self,globalThis,document,Document,Array,Object,String,Boolean,Math,Number,Symbol,Date,Function,Proxy,WeakMap,WeakSet,Set,Map,Reflect,Element,Node,RegExp,Error,TypeError,JSON,isNaN,parseFloat,parseInt,performance,console,decodeURI,encodeURI,decodeURIComponent,encodeURIComponent,navigator,undefined,location,history'
+export const HIJACK_LOCATION_KEYS = [
+  'host',
+  'hostname',
+  'port',
+  'protocol',
+  'origin',
+]