+++ src/sandbox/iframe/router.ts
@@ -12,7 +12,9 @@ import {
 import {
   assign,
   createURL,
+  rawDefineProperties,
 } from '../../libs/utils'
+import globalEnv from '../../libs/global_env'
 
 export function patchRouter (
   appName: string,
@@ -20,14 +22,30 @@ export function patchRouter (
   microAppWindow: microAppWindowType,
   browserHost: string,
 ): MicroLocation {
+  const rawHistory = globalEnv.rawWindow.history
   const childStaticLocation = createURL(url)
   const childHost = childStaticLocation.protocol + '//' + childStaticLocation.host
   const childFullPath = childStaticLocation.pathname + childStaticLocation.search + childStaticLocation.hash
 
   // rewrite microAppWindow.history
   const microHistory = microAppWindow.history
+  // save history.replaceState, it will be used in updateMicroLocation
   microAppWindow.rawReplaceState = microHistory.replaceState
+  // rewrite microAppWindow.history
   assign(microHistory, createMicroHistory(appName, microAppWindow.location))
+  // scrollRestoration proxy to rawHistory
+  rawDefineProperties(microHistory, {
+    scrollRestoration: {
+      configurable: true,
+      enumerable: true,
+      get () {
+        return rawHistory.scrollRestoration
+      },
+      set (value: string) {
+        rawHistory.scrollRestoration = value
+      }
+    }
+  })
 
   /**
    * Init microLocation before exec sandbox.start