+++ src/sandbox/router/index.ts
@@ -35,7 +35,7 @@ export {
 } from './event'
 export {
   getNoHashMicroPathFromURL,
-  getRouterMode,
+  initRouterMode,
   isRouterModeCustom,
   isRouterModeSearch,
 } from './core'
@@ -44,11 +44,6 @@ export {
   releasePatchHistory,
 } from './history'
 
-/**
- * TODO: 关于关闭虚拟路由系统的custom、history模式
- * 1. 是否需要发送popstate事件，为了减小对基座的影响，现在不发送
- * 2. 关闭后导致的vue3路由冲突问题需要在文档中明确指出(2处：在关闭虚拟路由系统的配置那里着重说明，在vue常见问题中说明)
- */
 /**
  * The router system has two operations: read and write
  * Read through location and write through history & location
@@ -78,6 +73,9 @@ export function initRouteStateWithURL (
   const microPath = getMicroPathFromURL(appName)
   if (microPath) {
     updateMicroLocation(appName, microPath, microLocation, 'auto')
+    if (isRouterModePure(appName)) {
+      removePathFromBrowser(appName)
+    }
   } else {
     updateBrowserURLWithLocation(appName, microLocation, defaultPage)
   }
@@ -101,10 +99,7 @@ export function updateBrowserURLWithLocation (
     attachRouteToBrowserURL(
       appName,
       setMicroPathToURL(appName, microLocation),
-      setMicroState(
-        appName,
-        null,
-      ),
+      setMicroState(appName, null, microLocation),
     )
   }
   // trigger guards after change browser URL
@@ -124,14 +119,13 @@ export function clearRouteStateFromURL (
   microLocation: MicroLocation,
   keepRouteState: boolean,
 ): void {
-  if (!isRouterModeCustom(appName)) {
-    if (!keepRouteState) {
-      const { pathname, search, hash } = createURL(url)
-      updateMicroLocation(appName, pathname + search + hash, microLocation, 'prevent')
-    }
-    if (!isRouterModePure(appName)) {
-      removePathFromBrowser(appName)
-    }
+  // TODO: keep-router-state 功能太弱，是否可以增加优先级，或者去掉
+  if (!keepRouteState && !isRouterModeCustom(appName)) {
+    const { pathname, search, hash } = createURL(url)
+    updateMicroLocation(appName, pathname + search + hash, microLocation, 'prevent')
+  }
+  if (!isRouterModePure(appName)) {
+    removePathFromBrowser(appName)
   }
 
   clearRouterWhenUnmount(appName)